Index products with the same data filtering as before
